Divide result reporter code into DAGResultReporter
